Move type coercion validation to a separate file
